+++ application-domain/src/main/kotlin/hs/kr/equus/application/domain/application/usecase/ApplicationPdfUseCase.kt
@@ -1,23 +1,21 @@
-package hs.kr.equus.application.global.util.pdf
+package hs.kr.equus.application.domain.application.usecase
 
 import hs.kr.equus.application.domain.application.exception.ApplicationExceptions
 import hs.kr.equus.application.domain.application.model.Application
+import hs.kr.equus.application.domain.application.spi.ApplicationPdfGeneratorPort
 import hs.kr.equus.application.domain.application.spi.QueryApplicationPort
 import hs.kr.equus.application.domain.graduationInfo.exception.GraduationInfoExceptions
-import hs.kr.equus.application.domain.score.exception.ScoreExceptions.ScoreNotFoundException
-import hs.kr.equus.application.domain.score.spi.ExistsScorePort
 import hs.kr.equus.application.domain.score.spi.QueryScorePort
+import hs.kr.equus.application.global.annotation.UseCase
 import hs.kr.equus.application.global.security.spi.SecurityPort
-import org.springframework.stereotype.Component
 
 
-@Component
+@UseCase
 class ApplicationPdfUseCase(
     private val securityPort: SecurityPort,
     private val queryApplicationPort: QueryApplicationPort,
-    private val existsScorePort: ExistsScorePort,
     private val queryScorePort: QueryScorePort,
-    private val applicationPdfGenerator: ApplicationPdfGenerator
+    private val applicationPdfGeneratorPort: ApplicationPdfGeneratorPort
 ) {
 
     fun getPreviewApplicationPdf(): ByteArray {
@@ -29,7 +27,7 @@ class ApplicationPdfUseCase(
 
         val calculatedScore = queryScorePort.queryScoreByReceiptCode(application.receiptCode)
 
-        return applicationPdfGenerator.generate(application, calculatedScore!!)
+        return applicationPdfGeneratorPort.generate(application, calculatedScore!!)
     }
 
     private fun validatePrintableApplicant(application: Application) {
