+++ application-infrastructure/src/main/kotlin/hs/kr/equus/application/domain/application/presentation/WepApplicationPdfAdapter.kt
@@ -1,6 +1,6 @@
 package hs.kr.equus.application.domain.application.presentation
 
-import hs.kr.equus.application.global.util.pdf.ApplicationPdfUseCase
+import hs.kr.equus.application.domain.application.usecase.ApplicationPdfUseCase
 import org.springframework.http.MediaType
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
